Added original AmesHousing dataset
We didn't want to use Kaggle's 50-50 train/test split so we added the
original dataset.

Changed the "Order" column in the original dataset to "Id" to match
current code, but the values are different. Anything identifying houses
on "Id" will not match what's on Kaggle.

For some reason Kaggle removed houses in the neighborhoods "Greens",
"GrnHill", and "Landmark". Keep an eye on this during modeling